src/requirements.txt
certifi
charset-normalizer
idna
pillow
requests
urllib3
tqdm
pyinstaller